Feat: Change wish, like api url
