+++ src/app.module.ts
@@ -10,6 +10,7 @@ import { AdminsModule } from './modules/admins/admins.module';
 import { AuthCodesModule } from './modules/auth-codes/auth-codes.module';
 import { AuthModule } from './modules/auth/auth.module';
 import { FiltersModule } from './modules/filters/filters.module';
+import { LikesModule } from './modules/likes/likes.module';
 import { PostsModule } from './modules/posts/posts.module';
 import { RanksModule } from './modules/ranks/ranks.module';
 import { RecommendationsModule } from './modules/recommendations/recommendations.module';
@@ -17,6 +18,7 @@ import { ReportPostsModule } from './modules/report-posts/report-posts.module';
 import { SpotsModule } from './modules/spots/spots.module';
 import { TasteThemesModule } from './modules/taste-themes/taste-themes.module';
 import { UsersModule } from './modules/users/users.module';
+import { WishesModule } from './modules/wishes/wishes.module';
 import { ViewController } from './views/view.controller';
 
 @Module({
@@ -36,6 +38,8 @@ import { ViewController } from './views/view.controller';
 		PostsModule,
 		AdminsModule,
 		ReportPostsModule,
+		WishesModule,
+		LikesModule,
 	],
 	controllers: [AppController, ViewController],
 	providers: [AppService],
