+++ src/modules/likes/dto/user-likes.dto.ts
@@ -1,7 +1,7 @@
 import { ApiProperty } from '@nestjs/swagger';
 import { IsArray, IsBoolean, IsDateString, IsNumber, IsString } from 'class-validator';
 import { LocationResponseDto } from 'src/modules/filters/dto/location-response.dto';
-import { CommentsUserResponseDto } from './comments-user-response.dto';
+import { CommentsUserResponseDto } from 'src/modules/users/dto/comments-user-response.dto';
 
 export class UserLikesDto {
 	@IsNumber()
