+++ src/modules/spots/spots.docs.ts
@@ -1,5 +1,5 @@
 import { applyDecorators } from '@nestjs/common';
-import { ApiBearerAuth, ApiBody, ApiConsumes, ApiOperation, ApiParam, ApiQuery, ApiResponse } from '@nestjs/swagger';
+import { ApiBearerAuth, ApiBody, ApiConsumes, ApiOperation, ApiQuery, ApiResponse } from '@nestjs/swagger';
 import { boolean } from 'joi';
 
 import { SwaggerMethodDoc } from 'src/swagger/swagger-method-doc-type';
@@ -117,28 +117,6 @@ export const ApiDocs: SwaggerMethodDoc<SpotsController> = {
 				description: '',
 				type: DetailSpotResponseDto,
 			}),
-		);
-	},
-	wishSpot(summary: string) {
-		return applyDecorators(
-			ApiOperation({
-				summary,
-				description: '여행지 찜하기',
-			}),
-			ApiParam({
-				name: 'spotId',
-				type: Number,
-			}),
-			ApiParam({
-				name: 'isWish',
-				type: Number,
-			}),
-			ApiResponse({
-				status: 201,
-				description: '',
-				type: Boolean,
-			}),
-			ApiBearerAuth('Authorization'),
 		);
 	},
 	getSnsPostUrls(summary: string) {
